requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-firestore
agentpy  # Placeholder if using Agent Development Kit-style patterns
